+++ src/httpGatewayTimeoutInterceptor/index.ts
@@ -1,0 +1,2 @@
+export * from './httpGatewayTimeoutInterceptor';
+export * from './httpGatewayTimeoutInterceptorOptions';